requirements.txt
torch>=2.0.0
matplotlib>=3.5.0
numpy>=1.21.0
PyYAML>=6.0
requests>=2.25.0
tqdm>=4.64.0